backend/requirements.txt
django==4.2
psycopg2
Pillow
djangorestframework
psycopg2-binary
django-environ
gunicorn
